Moving InputPatterns into the util module
The common module is primarily the Time and Entropy services.  With the goal being to eliminate those and replace them with consistgen.  Therefore this needs to be moved out of the way so the entire common package can be removed.
