+++ server-manage/src/main/java/com/unitvectory/serviceauthcentral/server/manage/controller/ClientResolver.java
@@ -23,7 +23,7 @@ import org.springframework.graphql.data.method.annotation.SchemaMapping;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Controller;
-import com.unitvectory.serviceauthcentral.common.InputPatterns;
+
 import com.unitvectory.serviceauthcentral.server.manage.dto.AuthorizationType;
 import com.unitvectory.serviceauthcentral.server.manage.dto.ClientManagementCapabilitiesType;
 import com.unitvectory.serviceauthcentral.server.manage.dto.ClientScopeType;
@@ -34,6 +34,7 @@ import com.unitvectory.serviceauthcentral.server.manage.dto.ResponseType;
 import com.unitvectory.serviceauthcentral.server.manage.mapper.RequestJwtMapper;
 import com.unitvectory.serviceauthcentral.server.manage.service.ClientService;
 import com.unitvectory.serviceauthcentral.server.manage.service.ManagementCapabilitiesService;
+import com.unitvectory.serviceauthcentral.util.InputPatterns;
 import com.unitvectory.serviceauthcentral.util.exception.BadRequestException;
 
 /**
