+++ util/src/main/java/com/unitvectory/serviceauthcentral/util/InputPatterns.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.unitvectory.serviceauthcentral.common;
+package com.unitvectory.serviceauthcentral.util;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
