+++ util/src/test/java/com/unitvectory/serviceauthcentral/util/InputPatternsTest.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.unitvectory.serviceauthcentral.common;
+package com.unitvectory.serviceauthcentral.util;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
